% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/criteoData.R
\name{criteoData}
\alias{criteoData}
\title{Pull Criteo Campaign Report Data}
\usage{
criteoData(authToken, appToken, jobID)
}
\arguments{
\item{authToken}{Object created by \code{link{doCriteoAuth}}}

\item{appToken}{App Token as character string.}

\item{jobID}{Object created by \code{link{scedCriteoReport}}}
}
\value{
Data
}
\description{
This function manages the complete data download process. The requested data will be returned as a data frame.
Wrapper for \code{link{jobStatus}}, \code{link{getCriteoDownloadURL}} and \code{link{getCriteoData}}.
}

