% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCriteoData}
\alias{getCriteoData}
\title{Get Campaign Report Data}
\usage{
getCriteoData(URL, jobID)
}
\arguments{
\item{URL}{Report download URL created by \code{\link{getCriteoDownloadURL}}}

\item{jobID}{jobID generated by \code{\link{scedCriteoReport}}}
}
\value{
Data
}
\description{
This function returns the requested data as a data.frame
}

