% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCriteoDownloadURL.R
\name{getCriteoDownloadURL}
\alias{getCriteoDownloadURL}
\title{Get Download Url}
\usage{
getCriteoDownloadURL(authToken, appToken, jobID)
}
\arguments{
\item{authToken}{Authentication token generated by \code{\link{doCriteoAuth}}}

\item{appToken}{Application Token}

\item{jobID}{jobID generated by \code{\link{scedCriteoReport}}}
}
\value{
Download URL
}
\description{
This function returns the download Url of the report.
}

