% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCriteoJobStatus.R
\name{getCriteoJobStatus}
\alias{getCriteoJobStatus}
\title{Get Job Status}
\usage{
getCriteoJobStatus(authToken, appToken, jobID)
}
\arguments{
\item{authToken}{Authentication token generated by \code{\link{doCriteoAuth}}}

\item{appToken}{Application Token}

\item{jobID}{jobID generated by \code{\link{scedCriteoReport}}}
}
\value{
Job status (Pending or InProgress or Completed or Failed)
}
\description{
This function return the status of the scheduled report generated by \code{\link{scedCriteoReport}}.
}

