% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scedCriteoReport.R
\name{scedCriteoReport}
\alias{scedCriteoReport}
\title{Campaign Report}
\usage{
scedCriteoReport(authToken, appToken, campaigns, metrics, start, end)
}
\arguments{
\item{authToken}{Authentication token generated by \code{\link{doCriteoAuth}}}

\item{appToken}{Application Token}

\item{campaigns}{Vector of campaign IDs. Example: campaigns = c('12345', '01235', '98765')}

\item{metrics}{Vector of metrics. Example: metrics = c('clicks','impressions','cost','sales')
available metrics: clicks, impressions, ctr, revcpc, ecpm, cost, sales, convRate, orderValue, salesPostView, convRatePostView, orderValuePostView, costOfSale, impressionWin, costPerOrder}

\item{start}{Start Date, format: "YYYY-mm-dd"}

\item{end}{End Date, format: "YYYY-mm-dd"}
}
\value{
Report job ID
}
\description{
This function generates the campaign report statement and schedules the report. The API returns a job ID, which later will be used to receive the data from the API.
}

