% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCriteoAuth.R
\name{doCriteoAuth}
\alias{doCriteoAuth}
\title{API Authentication}
\usage{
doCriteoAuth(user, password, company, app, version)
}
\arguments{
\item{user}{Username}

\item{password}{Password}

\item{company}{Company Name}

\item{app}{App Name of API project}

\item{version}{API version}
}
\value{
AuthToken
}
\description{
Initiate the Criteo API authentication process. The function returns the authentication token.
}

