% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCriteoAccount.R
\name{getCriteoAccount}
\alias{getCriteoAccount}
\title{Get Criteo Account Information}
\usage{
getCriteoAccount(authToken, appToken)
}
\arguments{
\item{authToken}{Authentication token generated by \code{\link{doCriteoAuth}}}

\item{appToken}{Application Token}
}
\value{
Account Information
}
\description{
This function returns the Criteo Account Information.
}

