% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCriteoCampaigns.R
\name{getCriteoCampaigns}
\alias{getCriteoCampaigns}
\title{Get Criteo Campaigns}
\usage{
getCriteoCampaigns(authToken, appToken)
}
\arguments{
\item{authToken}{Authentication token generated by \code{\link{doCriteoAuth}}}

\item{appToken}{Application Token}
}
\value{
Campaign Information
}
\description{
This function returns the Criteo Campaign Information.
}

