% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCriteoData.R
\name{getCriteoData}
\alias{getCriteoData}
\title{Get Campaign Report Data}
\usage{
getCriteoData(URL, jobID)
}
\arguments{
\item{URL}{Report download URL created by \code{\link{getCriteoDownloadURL}}}

\item{jobID}{jobID generated by \code{\link{scedCriteoReport}}}
}
\value{
Data
}
\description{
This function returns the requested data as a data frame.
}

